added exceptions

+++ Delivery_app/src/main/java/com/delivery/Delivery_app/service/CartService.java
@@ -2,6 +2,7 @@ package com.delivery.Delivery_app.service;
 
 import com.delivery.Delivery_app.dto.CartDTO;
 import com.delivery.Delivery_app.entity.Cart;
+import com.delivery.Delivery_app.exception.EmptyValueException;
 import com.delivery.Delivery_app.repository.CartRepository;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
@@ -19,6 +20,10 @@ public class CartService {
     //
     public void addFoodToCart(Long cartId, Long foodId, Long quantity) {
 
+        if(cartId == null || foodId == null || quantity == null){
+            throw new EmptyValueException("Input values are missing");
+        }
+
         int count = cartRepository.checkIfFoodInCart(cartId,foodId);
 
         if(count == 0) {
@@ -30,10 +35,16 @@ public class CartService {
     }
 
     public void deleteFoodFromCart(Long cartId, Long foodId) {
+        if(cartId == null || foodId == null){
+            throw new EmptyValueException("Input values are missing");
+        }
         cartRepository.deleteFoodFromCart(cartId,foodId);
     }
 
     public List getFoodListInCart(Long cartId) {
+        if(cartId == null){
+            throw new EmptyValueException("Cart ID is missing");
+        }
         List<Cart> foodList = cartRepository.getListOfFood(cartId);
 
         List<CartDTO> cartDTOList = modelMapper().map(foodList, new TypeToken<List<CartDTO>>() {}.getType());
